server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
pgvector==0.2.4
ollama==0.1.6
python-multipart==0.0.6
pydantic==2.5.3
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23